backend/requirements.txt
numpy<2
# For compatibility with Adobe evaluation scripts and sample scripts (chat_with_llm.py, generate_audio.py)
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1

# PDF processing
PyMuPDF==1.23.5

# Vector search and embeddings
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy>=1.26.0

# LLM for insights
google-generativeai==0.3.2

# Audio generation (Azure)
azure-cognitiveservices-speech==1.37.0

# Optional ML dependencies
scikit-learn==1.3.0

# Environment management
python-dotenv==1.0.0

# Additional utilities
typing-extensions==4.8.0
